wip/REST-3

+++ mrs_plugin/lib/general.py
@@ -1,4 +1,5 @@
 # Copyright (c) 2022, 2026, Oracle and/or its affiliates.
+# Copyright (c) 2026, MariaDB
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License, version 2.0,
@@ -206,7 +207,8 @@ def configure(
             else mysqlsh.globals.shell.open_session(session.connection_options)
         )
 
-        if edition is None or edition.lower() != "heatwave":
+        if session.server_vendor != "MariaDB" and (edition is None or edition.lower() != "heatwave"):
+
             # For now, let's remove any previous version of the mysql_tasks schema
             session.run_sql("DROP SCHEMA IF EXISTS mysql_tasks")
